Resolved gcc warnings.

+++ src/rlv12/rlv12command/rlv12command.cpp
@@ -2,14 +2,14 @@
 #include "rlv12/rlv12.h"
 
 RLV12Command::RLV12Command (int32_t trackNr, int32_t sectorNr,
-    int32_t memoryAddress, int32_t wordCount)
+    int32_t memoryAddress, size_t wordCount)
     : 
     trackNumber_ {trackNr},
     sectorNumber_ {sectorNr},
     memoryAddress_ {memoryAddress},
     wordCount_ {wordCount}
 {
-    int32_t maxWordCount = (RL_NUMSC - sectorNumber_) * RL_NUMWD;
+    size_t maxWordCount = (RL_NUMSC - sectorNumber_) * RL_NUMWD;
 
     // Track overrun?
     if (wordCount_ > maxWordCount)                                         
@@ -26,7 +26,7 @@ int32_t RLV12Command::sectorNumber() const
 int32_t RLV12Command::memoryAddress() const
     { return memoryAddress_; }
 
-int32_t RLV12Command::wordCount () const
+size_t RLV12Command::wordCount () const
     { return wordCount_; }
 
 // Return the position of the in this command specified track
