+++ src/rlv12/rlv12command/rlv12command.h
@@ -15,17 +15,17 @@ protected:
     int32_t trackNumber_;
     int32_t sectorNumber_;
     int32_t memoryAddress_;
-    int32_t wordCount_;
+    size_t wordCount_;
 
 public:
     RLV12Command (int32_t trackNr, int32_t sectorNr, 
-        int32_t memoryAddress, int32_t wordCount);
+        int32_t memoryAddress, size_t wordCount);
 
     // Accessors
     int32_t trackNumber() const;
     int32_t sectorNumber() const;
     int32_t memoryAddress() const;
-    int32_t wordCount () const;
+    size_t wordCount () const;
 
     // Return the position of the in this command specified track
     // as an offset in the file
